Working on UI and postMessage together
