+++ package.json
@@ -1,14 +1,35 @@
 {
-  "name": "Indian-User-Data-Tools",
+  "name": "Indian-User-Data-Generator",
   "version": "1.0.0",
   "description": "Generate sample data for Indian users",
-  "main": "code.min.js",
+  "main": "code.js",
   "scripts": {
-    "build": "tsc -p tsconfig.json"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "dev": "webpack --mode=development --watch",
+    "build": "webpack --mode=production"
   },
   "author": "Abhijit Chirde",
   "license": "MIT",
   "devDependencies": {
-    "@figma/plugin-typings": "^1.39.1"
+    "@figma/plugin-typings": "*",
+    "css-loader": "^6.8.1",
+    "html-inline-script-webpack-plugin": "^3.2.0",
+    "html-webpack-plugin": "^5.5.3",
+    "style-loader": "^3.3.3",
+    "ts-loader": "^9.4.3",
+    "typescript": "^5.1.3",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.88.0",
+    "webpack-cli": "^5.1.4"
+  },
+  "dependencies": {
+    "@types/react": "^18.2.14",
+    "@types/react-dom": "^18.2.6",
+    "@types/node": "^20.3.1",
+    "@types/styled-components": "^5.1.26",
+    "react": "^18.2.0",
+    "react-dev-utils": "^12.0.1",
+    "react-dom": "^18.2.0",
+    "styled-components": "^5.3.11"
   }
 }
