+++ src/components/BottomSection.tsx
@@ -1,0 +1,26 @@
+import * as React from "react";
+
+const BottomSection = () => {
+  const onDisclaimer = () => {
+    parent.postMessage({ pluginMessage: { type: "disclaimer" } }, "*");
+  };
+  return (
+    <div className="bottomContainer">
+      <button
+        className="disclaimer-text"
+        id="disclaimer"
+        onClick={onDisclaimer}
+      >
+        Disclaimer
+      </button>
+
+      <p className="credit-text">
+        <a href="https://abhijitchirde.com" target="_blank">
+          by Abhijit
+        </a>
+      </p>
+    </div>
+  );
+};
+
+export default BottomSection;
