+++ src/ui.css
@@ -1,0 +1,316 @@
+body {
+  background-color: #e2f7ff;
+  font-family: Arial, Helvetica, sans-serif;
+}
+
+/* tabs related styling */
+.tab {
+  overflow: hidden;
+  background-color: transparent;
+  width: 100%;
+  text-align: center;
+  margin-bottom: 0;
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 0.5rem;
+}
+
+.tab button {
+  border-bottom: 2px solid #fdfdfd;
+  border-top: none;
+  border-right: none;
+  border-left: none;
+  border-radius: 0.5rem 0.5rem 0 0;
+  outline: none;
+  cursor: pointer;
+  padding: 0.5rem 0.5rem;
+  margin-top: 0.5rem;
+  background-color: transparent;
+  text-align: center;
+  font-size: 0.8rem;
+  font-weight: 300;
+  color: rgb(158, 171, 196);
+  display: inline-block;
+  text-align: center;
+}
+
+.tab button:hover {
+  background-color: #fdfdfd;
+}
+
+.tab button.active {
+  background-color: #fdfdfd;
+  border-bottom: 2px solid rgba(137, 190, 255, 40);
+  /* border-radius: 1px solid #fcf7f9; */
+  color: rgb(79, 84, 107);
+  transition-duration: 0.4s;
+  font-weight: 300;
+}
+
+.tab button:disabled {
+  background: white;
+  color: black;
+  font-weight: 500;
+}
+
+.tabcontent {
+  display: block;
+  border-top: none;
+  align-items: center;
+  width: 100%;
+  margin: 0;
+  padding: 0;
+}
+
+/* content card styling */
+
+.tab-container {
+  margin: 0;
+  text-align: center;
+}
+
+/* Buttons styling */
+
+.buttons-list {
+  text-align: center;
+}
+
+.button-group {
+  text-align: center;
+  padding: 0 0.5rem 0.5rem;
+  margin-bottom: 1.5%;
+  background-color: #fff;
+  border: 1px solid #fff;
+  box-shadow: 0 1px 5px rgba(200, 200, 220, 0.3);
+  border-radius: 5px;
+}
+
+.group-label {
+  text-align: center;
+  font-family: Arial, Helvetica, sans-serif;
+  font-weight: lighter;
+  color: grey;
+  font-size: 0.8rem;
+  margin: 0.7rem 0;
+}
+
+.buttons-div {
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  gap: 0.6rem;
+  width: 100%;
+  align-items: center;
+}
+
+.button-pri {
+  font-size: 0.6rem;
+  font-weight: 400;
+  padding: 0.3rem;
+  min-width: 5.6rem;
+  height: 1.4rem;
+  width: auto;
+  border: 0.5px solid #c7c7c7;
+  background-color: #fff;
+  border-radius: 4px;
+  color: #6f6f70;
+  cursor: pointer;
+  /* box-shadow: 0 2px 3px rgba(19, 103, 230, 0.2); */
+}
+
+.button-pri:hover {
+  background-color: #1471fd; /* #396EB0 */
+  border: 0.5px solid #1471fd;
+  color: #fff;
+  transition-duration: 0.1s;
+}
+
+.button-pri:active {
+  background-color: #074aaf;
+  border: 0.5px solid #074aaf;
+  color: #fff;
+  transition-duration: 0.1s;
+  box-shadow: none;
+}
+
+.dropdown-label {
+  display: inline-block;
+  color: #6f6f70;
+  font-weight: 500;
+  font-size: 0.65rem;
+  padding-left: 3.7%;
+  padding-right: 3%;
+  vertical-align: middle;
+}
+
+.dropdown-main {
+  display: inline-block;
+  border-radius: 3px;
+  width: auto;
+  height: 1.4rem;
+  font-size: 0.65rem;
+  font-weight: 500;
+  text-align: left;
+  padding-left: 1%;
+  border: 0.5px solid #c7c7c7;
+  color: #6f6f70;
+  margin-right: 3.2%;
+  vertical-align: middle;
+}
+
+.dropdown-main:disabled {
+  background-color: rgb(235, 235, 235);
+}
+
+.dropdown-item {
+  font-size: 0.65rem;
+  height: 16%;
+  padding: 1% 1%;
+  color: #6f6f70;
+}
+
+/* checkbox elements related styling */
+
+.checkbox-list {
+  text-align: center;
+}
+
+.checkbox-group {
+  text-align: center;
+  padding-bottom: 0%;
+  margin-bottom: 1.5%;
+  background-color: #fff;
+  border: 1px solid #fff;
+  box-shadow: 0 1px 5px rgba(200, 200, 220, 0.3);
+  border-radius: 5px;
+}
+
+.checkboxes {
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  gap: 0.6rem;
+  width: 100%;
+  align-items: center;
+}
+
+.checkbox-div {
+  padding: 0 0.3rem;
+  min-width: 5.6rem;
+  height: 1.4rem;
+  width: auto;
+  display: inline-block;
+  text-align: left;
+  vertical-align: middle;
+}
+
+.checkbox {
+  margin: 1% 1%;
+  vertical-align: middle;
+}
+
+.checkbox-label {
+  font-family: Arial, Helvetica, sans-serif;
+  font-size: 0.65rem;
+  font-weight: 400;
+  margin-left: 2%;
+  color: #6f6f70;
+  vertical-align: middle;
+}
+
+/* Generate table button area */
+
+.separate-button-div {
+  display: block;
+  width: auto;
+  padding-top: 1.5%;
+  padding-bottom: 0%;
+}
+
+.user-input {
+  display: inline-block;
+  height: 4.4%;
+  width: 10%;
+  font-size: 0.8rem;
+  border: 1px solid #adadad;
+  border-radius: 3px;
+  color: #747474;
+  margin-left: 2%;
+  margin-right: 3.2%;
+  text-align: center;
+}
+
+.input-label {
+  display: inline-block;
+  font-family: Arial, Helvetica, sans-serif;
+  font-size: 0.7rem;
+  color: #6f6f70;
+}
+
+.button-large {
+  font-size: 0.8rem;
+  font-weight: 400;
+  margin: 1.2% 0.8%;
+  padding: 1.2% 0.2%;
+  width: 32.56%;
+  display: inline;
+  border: 0.5px solid #6f6f70;
+  background-color: #fdfdfd;
+  border-radius: 4px;
+  color: #6f6f70;
+  cursor: pointer;
+}
+
+.button-large:hover {
+  background-color: #1471fd;
+  border: 0.5px solid #1471fd;
+  color: #fff;
+  transition-duration: 0.2s;
+}
+
+.button-large:active {
+  background-color: #074aaf;
+  border: 0.5px solid #0b5ed7;
+  color: #fff;
+  transition-duration: 0.2s;
+}
+
+/* Footer styling */
+
+.bottomContainer {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  padding: 0 1rem;
+}
+
+.disclaimer-text {
+  border: none;
+  font-size: 0.6rem;
+  background: none;
+  color: rgb(158, 171, 196);
+  cursor: pointer;
+  display: inline-block;
+  padding: 0;
+}
+
+.disclaimer-text:hover {
+  color: rgb(108, 125, 155);
+  transition-duration: 0.3s;
+}
+
+.credit-text {
+  font-size: 0.6rem;
+  color: rgb(158, 171, 196);
+  display: inline-block;
+}
+
+a:link,
+a:visited {
+  color: rgb(158, 171, 196);
+  cursor: pointer;
+  text-decoration: none;
+}
+
+a:hover {
+  color: rgb(108, 125, 155);
+  transition-duration: 0.3s;
+}
