+++ src/ui.html
@@ -1,0 +1,1 @@
+<div id="react-page"></div>
