+++ tsconfig.json
@@ -1,10 +1,8 @@
 {
   "compilerOptions": {
     "target": "es6",
-    "lib": ["es6"],
-    "typeRoots": [
-      "./node_modules/@types",
-      "./node_modules/@figma"
-    ]
-  }
+    "jsx": "react",
+    "typeRoots": ["./node_modules/@types", "./node_modules/@figma"]
+  },
+  "include": ["src/**/*.ts", "src/**/*.tsx"]
 }
